Refactoring film card component + hook
