+++ src/components/cards/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Template } from '../../components/template'
+import { Template } from '../template'
 import { useCards } from './hooks'
 import { makeStyles } from '@material-ui/core/styles'
 import Card from '@material-ui/core/Card'
