+++ src/navigation/routes.js
@@ -1,8 +1,8 @@
-import { Menu } from '../screens/Section1'
+import { Home } from '../screens/Home/home'
 
 export const routes = [
   {
     path: '/',
-    component: Menu
+    component: Home
   }
 ]
